+++ README.md
@@ -1,5 +1,5 @@
-#Topdoc Utils [![Build Status](https://travis-ci.org/topcoat/topdoc-utils.png)](https://travis-ci.org/topcoat/topdoc-utils)
+#Taborlin Utils [![Build Status](https://travis-ci.org/topcoat/Taborlin-utils.png)](https://travis-ci.org/topcoat/Taborlin-utils)
 
-[![NPM](https://nodei.co/npm/topdoc-utils.png)](https://nodei.co/npm/topdoc-utils)
+[![NPM](https://nodei.co/npm/Taborlin-utils.png)](https://nodei.co/npm/Taborlin-utils)
 
-Small utility functions used by [Topdoc](http://github.com/topcoat/topdoc) and [Topdoc Parse](http://github.com/topcoat/topdoc-parse)
+Small utility functions used by [Taborlin](https://github.com/Taborlin/Taborlin) and [Taborlin Parse](https://github.com/Taborlin/taborlin-parse)
